Upgrade dependencies versions and enabled R8
The only one soulution for issue #8 is to enable R8.
I could have also set 'android.proguard.enableRulesExtraction=false',
but it is not documented, it is experimental, and I do not know if it
disable the extraction of the rules only from JAR modules or from AAR
modules, too.
Anyway R8 will be enabled by default in the next version of Android
Studio, so no reasons to struggle for other solutions.

Enabling R8 also solves issue #7.